Convert derived class headers to cpp
