+++ libplot/IEventDisplay.cpp
