+++ libplot/SelectionEfficiencyPlot.cpp
