+++ libplot/StackedHistogramPlot.cpp
